packages/sdk-python/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "verum"
version = "0.1.0"
description = "Verum SDK — connect your AI service to The Verum Loop"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.28.1,<1.0",
    "pydantic>=2.7,<3.0",
]

[project.urls]
Homepage = "https://github.com/xzawed/verum"
Documentation = "https://docs.verum.dev"

[project.optional-dependencies]
instrument = [
    "openinference-instrumentation-openai>=0.1.0",
    "opentelemetry-exporter-otlp-proto-http>=1.0",
    "opentelemetry-sdk>=1.0",
]
dev = [
    "pytest>=8",
    "pytest-asyncio>=0.23",
    "pytest-cov>=7.1.0",
    "respx>=0.23.1",
    "freezegun>=1.5",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.coverage.run]
source = ["src"]
relative_files = true

[tool.coverage.report]
exclude_lines = ["if __name__ == .__main__.:"]
